a lot of yapping

+++ src/main/java/mod/maxbogomol/wizards_reborn/common/item/equipment/WissenWandItem.java
@@ -374,7 +374,7 @@ public class WissenWandItem extends Item {
             gui.blit(barTexture, x, y, 0, 0, 48, 10, 64, 64);
             int width = 32;
             double v = (double) maxValue / (double) value;
-            width /= (int) v;
+            width /= v;
             if (width == 0 && v > 0 && !Double.isInfinite(v)) width = 1;
             gui.blit(barTexture, x + 8, y + 1, 0, 10, width, 8, 64, 64);
         }
